services/contract-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-docx==1.1.2
python-multipart==0.0.19
requests==2.32.3
PyPDF2==3.0.1
Pillow==11.0.0
reportlab==4.2.5

